requirements.txt
absl-py==1.0.0
anykeystore==0.2
apex==0.9.10.dev0
astunparse==1.6.3
attrs==21.4.0
azure-devtools==1.2.0
black==22.3.0
cachetools==5.0.0
cblack==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
chex==0.1.3
click==8.1.2
clu==0.0.6
ConfigArgParse==1.5.3
contextlib2==21.6.0
cryptacular==1.6.2
cycler==0.11.0
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
dm-tree==0.1.7
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.3
filelock==3.6.0
flake8==3.7.9
flaky==3.7.0
flatbuffers==2.0
flax==0.4.1
fonttools==4.33.3
functorch==0.1.1
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
hupper==1.10.3
identify==2.4.12
idna==3.3
immutabledict==2.2.1
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
jax==0.3.7
jaxlib==0.3.7
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
ml-collections==0.1.1
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
optax==0.1.2
ott-jax==0.2.5
packaging==21.3
pandas==1.4.2
PasteDeploy==2.1.1
pathspec==0.9.0
pathtools==0.1.2
pbkdf2==1.3
Pillow==9.1.0
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.18.1
promise==2.3
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==3.0.8
pyramid==2.0
pyramid-mailer==0.15.1
pytest==7.1.2
pytest-flake8==1.1.0
pytest-rerunfailures==10.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
ratelimit==2.2.1
repoze.sendmail==4.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentencepiece==0.1.96
sentry-sdk==1.6.0
setproctitle==1.2.3
setuptools-scm==6.4.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.38
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.7.0
tensorflow-text==2.8.2
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
timm==0.3.2
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.8.0
torchvision==0.8.2
tqdm==4.64.0
transaction==3.0.1
translationstring==1.4
typing_extensions==4.2.0
urllib3==1.26.9
vcrpy==4.1.1
velruse==1.1.1
venusian==3.0.0
virtualenv==20.14.1
wandb==0.12.19
WebOb==1.8.7
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yarl==1.7.2
zipp==3.8.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
